requirements.txt
absl-py==2.3.1
anyio==4.11.0
asgiref==3.9.2
astunparse==1.6.3
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
firebase_admin==7.1.0
Flask==3.1.2
flatbuffers==25.9.23
gast==0.6.0
gcloud==0.18.3
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
json5==0.12.1
jwcrypto==1.5.6
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
numpy==2.3.5
oauth2client==4.1.3
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pickle-mixin==1.0.2
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==2.0.1
